+++ ontology/abstract/__init__.py
@@ -1,5 +1,0 @@
-from .language_evolution import LanguageEvolutionProcess
-from .science_paradigms import ScienceParadigmsProcess
-from .ai_emergence import AIEmergenceProcess
-
-__all__ = ['LanguageEvolutionProcess', 'ScienceParadigmsProcess', 'AIEmergenceProcess']