+++ ontology/biology_medicine/__init__.py
@@ -1,6 +1,0 @@
-from .metabolism import MetabolismProcess
-from .immune_system import ImmuneSystemProcess
-from .population_genetics import PopulationGeneticsProcess
-from .oncology import OncologyProcess
-
-__all__ = ['MetabolismProcess', 'ImmuneSystemProcess', 'PopulationGeneticsProcess', 'OncologyProcess']