+++ ontology/origins_of_life/__init__.py
@@ -1,3 +1,0 @@
-from .tri_regime_bioenergetics import TriRegimeBioenergeticsProcess
-
-__all__ = ['TriRegimeBioenergeticsProcess']