+++ ontology/physics/__init__.py
@@ -1,6 +1,0 @@
-from .plasma_dynamics import PlasmaDynamicsProcess
-from .nuclear_dynamics import NuclearDynamicsProcess
-from .torus_knots import TorusKnotProcess
-from .quantum_vacuum import QuantumVacuumProcess
-
-__all__ = ['PlasmaDynamicsProcess', 'NuclearDynamicsProcess', 'TorusKnotProcess', 'QuantumVacuumProcess']