+++ ontology/socioeconomic/__init__.py
@@ -1,9 +1,0 @@
-from .financial_markets import FinancialMarketProcess
-from .education import EducationProcess
-from .trade_networks import TradeNetworksProcess
-from .governance import GovernanceProcess
-from .crises import CrisesProcess
-from .migration import MigrationProcess
-from .urban_systems import UrbanSystemsProcess
-
-__all__ = ['FinancialMarketProcess', 'EducationProcess', 'TradeNetworksProcess', 'GovernanceProcess', 'CrisesProcess', 'MigrationProcess', 'UrbanSystemsProcess']